requirements.txt
flask
flask-cors
python-slugify
requests
